admin user coding

+++ app/admin/admin_users.rb
@@ -1,15 +1,14 @@
 ActiveAdmin.register AdminUser do
-  permit_params :first_name, :last_name, :avatar_url, :birthday, :phone_number, :password, :password_confirmation, :permission, :avatar
+  permit_params :first_name, :last_name, :avatar_url, :birthday, :phone_number, :email, :password, :password_confirmation, :permission, :avatar
 
   index do
     selectable_column
     id_column
     column :first_name
     column :last_name
-    column :avatar_url
-    # do |avatar_url|
-    #   link_to(image_tag(avatar_url), avatar_url)
-    # end
+    column 'Avatar', :avatar_url do |item|
+      link_to(image_tag(item.avatar_url, width: '70px', height: '70px'), item.avatar_url) unless item.avatar_url.nil?
+    end
     column(:birthday) { |time| time.birthday.strftime('%B %d, %Y') }
     column :phone_number
     column :email
@@ -35,10 +34,13 @@ ActiveAdmin.register AdminUser do
       f.input :first_name, as: :string
       f.input :last_name, as: :string
       f.input :avatar, as: :file
-      # , hint: f.template.image_tag(f.object.avatar_url)
       f.input :birthday, as: :datepicker, input_html: { value: f.object.birthday.try(:strftime, '%Y-%m-%d') }
       f.input :phone_number, input_html: { maxlength: 10, oninput: "this.value = this.value.replace(/[^0-9]/g, '').replace(/(\\..*?)\\..*/g, '$1');" }
-      f.input :email, input_html: { readonly: true, disabled: true}
+      if f.object.new_record?
+        f.input :email
+      else
+        f.input :email, input_html: { readonly: true, disabled: true}
+      end
       f.input :password
       f.input :password_confirmation
       f.input :permission
@@ -53,16 +55,34 @@ ActiveAdmin.register AdminUser do
 
   controller do
     def create
+      unless params[:admin_user][:avatar].nil?
+        File.open(Rails.root.join('public', 'uploads', params[:admin_user][:avatar].original_filename), 'wb') do |file|
+          file.write(params[:admin_user][:avatar].read)
+        end
+        image_url = "public/uploads/#{params[:admin_user][:avatar].original_filename}"
+        @admin_user.avatar_url = Cloudinary::Uploader.upload(image_url)['url']
+        File.delete(image_url) if File.exist?(image_url)
+      end
+      params[:admin_user].reject! { |p| p['avatar'] }
+      super
     end
 
     def update
+      params[:admin_user].reject! { |p| p['email'] }
       @admin_user = AdminUser.find(params[:id])
-      File.open(Rails.root.join('public', 'uploads', params[:admin_user][:avatar].original_filename), 'wb') do |file|
-        file.write(params[:admin_user][:avatar].read)
+      unless @admin_user.avatar_url.nil?
+        public_id = @admin_user.avatar_url.split('/')[-1].split('.')[0]
+        Cloudinary::Uploader.destroy(public_id)
+        @admin_user.avatar_url = nil
       end
-      image_url = "public/uploads/#{params[:admin_user][:avatar].original_filename}"
-      @admin_user.avatar_url = Cloudinary::Uploader.upload(image_url)['url']
-      File.delete(image_url) if File.exist?(image_url)
+      unless params[:admin_user][:avatar].nil?
+        File.open(Rails.root.join('public', 'uploads', params[:admin_user][:avatar].original_filename), 'wb') do |file|
+          file.write(params[:admin_user][:avatar].read)
+        end
+        image_url = "public/uploads/#{params[:admin_user][:avatar].original_filename}"
+        @admin_user.avatar_url = Cloudinary::Uploader.upload(image_url)['url']
+        File.delete(image_url) if File.exist?(image_url)
+      end
       params[:admin_user].reject! { |p| p['avatar'] }
       super
     end
